+++ db/migrate/20141129004858_create_results.rb
@@ -1,7 +1,7 @@
 class CreateResults < ActiveRecord::Migration
   def change
     create_table :results do |t|
-      t.integer :value, default: 0
+      t.float :value, default: 0.0
       t.integer :ratee_id, index: true, null: false
       t.integer :rater_id, index: true, null: false
       t.references :evaluation, index: true, null: false
